Move logic out of DrawView
