+++ src/components/recording/StreamRenderer.js
